Place importing logic coupled to SQLAlchemy in its own file.
